Add FindLinks api to non-webdrivers
