+++ splinter/driver/lxmldriver.py
@@ -14,6 +14,7 @@ import lxml.etree
 import lxml.html
 from lxml.cssselect import CSSSelector
 from splinter.driver import DriverAPI, ElementAPI
+from splinter.driver.find_links import FindLinks
 from splinter.driver.element_present import ElementPresentMixIn
 from splinter.driver.xpath_utils import _concat_xpath_from_str
 from splinter.element_list import ElementList
@@ -29,6 +30,8 @@ class LxmlDriver(ElementPresentMixIn, DriverAPI):
         self._history = []
         self._last_urls = []
         self._forms = {}
+
+        self.links = FindLinks(self)
 
     def __enter__(self):
         return self
@@ -148,10 +151,10 @@ class LxmlDriver(ElementPresentMixIn, DriverAPI):
     def find_by_css(self, selector):
         xpath = CSSSelector(selector).path
         return self.find_by_xpath(
-            xpath, original_find="css", original_selector=selector
+            xpath, original_find="css", original_query=selector
         )
 
-    def find_by_xpath(self, xpath, original_find=None, original_selector=None):
+    def find_by_xpath(self, xpath, original_find=None, original_query=None):
         html = self.htmltree
 
         elements = []
@@ -165,7 +168,7 @@ class LxmlDriver(ElementPresentMixIn, DriverAPI):
                 elements.append((LxmlElement, xpath_element))
 
         find_by = original_find or "xpath"
-        query = original_selector or xpath
+        query = original_query or xpath
 
         return ElementList(
             [element_class(element, self) for element_class, element in elements],
@@ -175,12 +178,12 @@ class LxmlDriver(ElementPresentMixIn, DriverAPI):
 
     def find_by_tag(self, tag):
         return self.find_by_xpath(
-            "//%s" % tag, original_find="tag", original_selector=tag
+            "//%s" % tag, original_find="tag", original_query=tag
         )
 
     def find_by_value(self, value):
         elem = self.find_by_xpath(
-            '//*[@value="%s"]' % value, original_find="value", original_selector=value
+            '//*[@value="%s"]' % value, original_find="value", original_query=value
         )
         if elem:
             return elem
@@ -191,14 +194,14 @@ class LxmlDriver(ElementPresentMixIn, DriverAPI):
         return self.find_by_xpath(
             xpath_str,
             original_find="text",
-            original_selector=text,
+            original_query=text,
         )
 
     def find_by_id(self, id_value):
         return self.find_by_xpath(
             '//*[@id="%s"][1]' % id_value,
             original_find="id",
-            original_selector=id_value,
+            original_query=id_value,
         )
 
     def find_by_name(self, name):
